V1.2 Listed in README.md

+++ backend/controllers/CollectController.php
@@ -4,16 +4,16 @@ namespace backend\controllers;
 
 use Yii;
 use yii\filters\AccessControl;
-use common\models\Buy;
 use common\models\Storage;
+use common\models\Bookmark;
+use common\models\Add;
 use backend\models\Stock;
-use common\models\History;
+use backend\models\StorageSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use common\components\yii2GridViewState\widgets\GridView;
-
-use common\models\Bookmark;
+use backend\models\CollectSearch;
 
 // use backend\models\FileUploadForm;
 // use yii\web\UploadedFile;
@@ -21,9 +21,9 @@ use common\models\Bookmark;
 
 
 /**
- * SystemUserController implements the CRUD actions for SystemUser model.
+ * StorageController implements the CRUD actions for Storage model.
  */
-class SearchController extends Controller
+class CollectController extends Controller
 {
     /**
      * @inheritdoc
@@ -50,40 +50,43 @@ class SearchController extends Controller
     }
 
     /**
-     * Lists all SystemUser models.
+     * Lists all Storage models.
      * @return mixed
      */
     public function actionIndex()
     {
-        $model = new Buy();
+        $model = new Add();
 
-        $search = null;
-        $bookmarked = false;
-        $stock = Yii::$app->getRequest()->getQueryParam('Buy');
+        $searchModel = new CollectSearch();
+		$dataProvider = GridView::search($searchModel);
 
+        $stock_type_list = Stock::GetStockTypeList();
 
-        if($stock){
-            $stock = $stock['code'];
-            $search = str_pad($stock,5,'0',STR_PAD_LEFT);
-
-            $bookmark =
-                Bookmark::find()
-                    ->where(['user_id' => Yii::$app->user->identity->id])
-                    ->andWhere(['stock_code' => $search])
-                    ->asArray()->one();
-            if($bookmark != null) {
-                $bookmarked = true;
-            }
-        }
-
-
+        
         return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+            'stock_type_list' => $stock_type_list,
             'model' => $model,
-            'search' => $search,
-            'bookmarked' => $bookmarked,
         ]);
     }
 
+
+    public function actionGetstockinfo(){
+        \Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
+        $data = Yii::$app->request->post();
+
+        $stock = $data['stock'];
+        $stock = str_pad($stock,5,'0',STR_PAD_LEFT);
+
+        $price = Stock::GetStockPrice($stock);
+        $name = Stock::GetStockInfo($stock)['name'];
+        $lotsize = Stock::GetStockInfo($stock)['lotsize'];
+        $uaCode = Stock::GetStockInfo($stock)['uaCode'];
+
+
+        return ['stock' => $stock ,'price' => $price, 'name' => $name, 'lotsize' => $lotsize, 'uaCode' => $uaCode];
+    }
 
     public function actionGetbookmark(){
         \Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
@@ -92,11 +95,15 @@ class SearchController extends Controller
         $stock = $data['stock'];
         $stock = str_pad($stock,5,'0',STR_PAD_LEFT);
 
+        $price = Stock::GetStockPrice($stock);
+
         $bookmark =
             Bookmark::find()
                 ->where(['user_id' => Yii::$app->user->identity->id])
                 ->andWhere(['stock_code' => $stock])
                 ->one();
+
+        $db_id = null;
 
         if($bookmark == null) {
             $model_bookmark = new Bookmark();
@@ -105,20 +112,40 @@ class SearchController extends Controller
             $model_bookmark->stock_string = Stock::GetStockInfo($stock)['name'];
             $model_bookmark->save();
 
+            $db_id = $model_bookmark->getPrimaryKey();
+
             $result = 'success';
         } else {
             $result = 'fail';
-            $bookmark->delete();
         }
 
         $stock_name = Stock::GetStockInfo($stock)['name'];
 
 
-        return ['result' => $result, 'stock' => $stock, 'stock_name' => $stock_name];
+        return ['db_id' => $db_id,'result' => $result, 'stock' => $stock, 'stock_name' => $stock_name, 'price' => $price];
     }
 
+    public function actionDelbookmark(){
+        \Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
+        $data = Yii::$app->request->post();
+
+        $id = $data['id'];
+
+        $bookmark =
+            Bookmark::find()
+                ->where(['id' => $id])
+                ->one();
+
+        if($bookmark != null) {
+            $result = 'deleted';
+            $bookmark->delete();
+        } else {
+            $result = 'error';
+        }
 
 
+        return ['result' => $result, 'id' => $id];
+    }
 
 
     protected function findModel($id)
